+++ source/libs/uti/test_sge_lock_simple.c
@@ -93,7 +93,8 @@ static void *thread_function(void *anArg)
    SGE_LOCK(LOCK_GLOBAL, LOCK_READ);
 
 #if 1
-   DPRINTF(("Thread %u sleeping at %d\n", sge_locker_id(), sge_get_gmt()));
+   DPRINTF(("Thread %u sleeping at %d\n",(unsigned int) sge_locker_id(),
+            (int) sge_get_gmt()));
 #endif
    sleep(5);
